Updates img_proc_widget.py
Pop up warnings implemented for the case that models or folders are not selected, when a selection is required. Manually processed masks can be saved with individual names. Masks can be loaded and displayed with the corresponding images in the viewer. Model download options depend on the model URL. This feature needs to be extended and tested, if models are to be located on other places than Github.